se corrigio los Scripts de SQL,se corrigio el agregar alumno,esta listo agregar curso,eliminar curso,consultar curso,eliminar alumno,falta agregar profesor,eliminar profesor y conusltar profesor,y toda la parte de modificar y el modulo de examenes

+++ BD/Oracle/Procedimientos Almacenados/SP_AGREGAR_CURSO.sql
@@ -3,7 +3,7 @@ PROCEDURE SP_AGREGAR_CURSO
 (
   p_nombre varchar2,
   p_ciclo varchar2,
-  p_creditos integer,
+  p_creditos float,
   p_usuario varchar2
 )
 AS
